refactor text test
